+++ CHANGELOG.md
@@ -4,7 +4,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
-
+* Mysql Adapter
 
 # [0.3.0] 2016-12-17
 * Check for result.rows.size in queries - [@neovintage](https://github.com/neovintage)